CustomVision/flask_app/requirements.txt
Flask
azure-cognitiveservices-vision-customvision